Update for graphics

+++ src/libs/boot/bootparams.h
@@ -61,7 +61,7 @@ typedef struct {
 	uint8_t reserved_position;
 	uint8_t direct_color_attributes;
  
-	uint32_t framebuffer;		// physical address of the linear frame buffer; write here to draw to the screen
+	uintptr_t framebuffer;		// physical address of the linear frame buffer; write here to draw to the screen
 	uint32_t off_screen_mem_off;
 	uint16_t off_screen_mem_size;	// size of memory in the framebuffer but not being displayed on the screen
 	uint8_t reserved1[206];
